Refactor: proxy now uses its own Link class.
