+++ lightnion/link.py
@@ -1,406 +1,220 @@
 import logging
 import socket
+import queue
 import ssl
-import asyncio
 
 import lightnion as lnn
-import lightnion.cell
-import lightnion.utils
-from lightnion.proxy import fake_circuit_id
 
+class link:
+    """An established Tor link, send and receive messages in separate threads.
 
-class InvalidCellHeaderException(Exception):
-    pass
+    :param io: socket.io instance that wraps the TLS/SSLv3 connection
+    :param int version: link version
 
-class InvalidAuthCellException(Exception):
-    pass
+    Usage::
 
-class InvalidCertsCellException(Exception):
-    pass
+      >>> import lightnion as lnn
+      >>> link = lnn.link.initiate('127.0.0.1', 5000)
+      >>> link.version
+      5
+      >>> link.send(lnn.cell.create_fast.pack(2**31))
+      >>> lnn.cell.created_fast.cell(link.get(circuit_id=2**31)).valid
+      True
+      >>> link.close()
+    """
+    def __init__(self, io, version, circuits=[0], max_queue=2048):
+        self.max_queue = 2048
+        self.version = version
+        self.last_id = 0
+        self.io = io
 
-class InvalidDestroyCellException(Exception):
-    pass
+        self.circuits = dict()
+        self.register(lnn.create.circuit(0, None))
 
-class InvalidNetInfoCellException(Exception):
-    pass
+    def pull(self, block=True):
+        try:
+            payload = self.io.recv(block=block)
+        except queue.Empty:
+            return False
 
-class InvalidVersionCellException(Exception):
-    pass
+        # We know that receiver.get() will give you a cell with a well-formed
+        # header, thus we do not validate it one more time.
+        #
+        # We doesn't handle VERSIONS cells with shorter circuit_id.
+        #
+        header = lnn.cell.header(payload)
+        if not header.circuit_id in self.circuits:
+            raise RuntimeError('Got circuit {} outside {}, cell: {}'.format(
+                header.circuit_id, list(self.circuits), payload))
 
-class NoSupportedVersionException(Exception):
-    pass
-
-class Link:
-    def __init__(self, guard, versions=(4,5)):
-        """
-        Handler for communications between the proxy and a guard relay.
-        :param guard: guard tor relay with with to establish a link.
-        :param versions: versions supported by the proxy
-        """
-
-        host = guard['router']['address']
-        port = guard['router']['orport']
-
-        # Queue containing cells to be send to the tor relay.
-        self.to_send = asyncio.Queue(16384)
-
-        # Buffer containing beginning of potential imcomplete cell.
-        self.buffer = b''
-
-        # The link is bound to a specific guard node.
-        self.guard = guard
-
-        logging.warning('Guard {}'.format(guard['router']['nickname']))
-
-        # The first bit of the circuit id must be 1.
-        self.circuit_id = 0x80000000
-
-        ctxt = ssl.SSLContext(ssl.PROTOCOL_TLS)
-
-        # TLS 1.3 disabled for compatibility issues.
-        # https://trac.torproject.org/projects/tor/ticket/28616
-        ctxt.options |= ssl.OP_NO_TLSv1_3
-
-        # channel manager set later.
-        self.channel_manager = None
-
-        # The connection to the tor relay.
-        self.connection = self._handler(host, port, ctxt, versions)
-
-        self.cell_sent = 0
-        self.cell_recv = 0
-
-        logging.debug('Link: Link prepared ({} : {})'.format(host, port))
-
-
-    def set_channel_manager(self, channel_manager):
-        """
-        Set the channel manager to which the traffic should be send.
-        :param channel_manager: channel manager to use.
-        """
-
-        self.channel_manager = channel_manager
-
-        logging.debug('Link: Channel manager set.')
-
-
-    def gen_cid(self):
-        """
-        Generate a new circuit id.
-        :return: new circuit id
-        """
-        self.circuit_id += 1
-
-        logging.debug('Link: New circuit id generated {}'.format(self.circuit_id))
-
-        return self.circuit_id
-
-
-    async def schedule_to_send(self, cell_raw, channel):
-        """
-        Coroutine
-        Order some data to be sent to the tor relay.
-        :param cell: cell to be sent.
-        :param channel: channel from which the cell is sent.
-        """
-
-        # Set correct circuit id.
-        cell = lnn.cell.header_view.write(cell_raw, circuit_id=channel.cid)
-        #cell = lnn.utils.cell_set_cid(cell_raw, channel.cid)
-
-        header = lnn.cell.header(cell)
-        #cmd = lnn.utils.cell_get_cmd(cell)
-        #if cmd == 3:
-        #    cell = lnn.utils.cell_pad_rnd(cell)
-        #else:
-        #    cell = lnn.utils.cell_pad_null(cell)
-
-        #if cmd == 4:
+        # TODO: property handle DESTROY cells
+        circuit = self.circuits[header.circuit_id]
         if header.cmd is lnn.cell.cmd.DESTROY:
-            logging.debug('Link: channel {} asks for the circuit to be destroyed.'.format(channel.cid))
-            cell_validation = lnn.cell.destroy.cell(cell)
-            #if not lnn.utils.cell_is_valid(cell):
-            if not cell_validation.valid:
-                logging.debug('Link: Channel {} attempted to send an invalid cell.'.format(channel.cid))
-                return
-                #raise InvalidDestroyCellException()
-
-            # Mark the channel as destroyed
-            if not channel.destroyed.is_set():
-                channel.destroyed.set()
-
-        cell_padded = lnn.cell.pad(cell)
-
-        await self.to_send.put(cell_padded)
-        #await self.to_send.put(cell)
-        logging.debug('Link: Scheduled data from channel {} to be send.'.format(channel.cid))
-
-
-    async def _handle_tor_cmd_cell(self, cell):
-        logging.debug('Link: Handle management cell from tor relay.')
-
-
-    async def _recv(self, reader):
-        """
-        Recieve data from the tor relay and give it to the channel manager.
-        :param reader: asyncio StreamReader
-        """
-
-        while not reader.at_eof():
-            data = await reader.read(65536)
-
-            data = self.buffer + data
-
-            data_initial = data
-
-            # Let the channel manager do the multiplexing.
-            #logging.debug('Link: Received data\n{}'.format(data))
-
-            (cell, data) = lnn.utils.cell_slice_old(data)
-
-            data_cells = b''
-            cells = []
-
-            while cell is not None:
-
-                data_cells += cell
-                cells.append(cell)
-
-                logging.debug('Link: Spliced a cell. {}... {} bytes.'.format(cell[:20], len(cell)))
-                # Analyse header to select correct channel.
-                header = lnn.cell.header(cell)
-                cid = header.circuit_id
-                #cid = lnn.utils.cell_get_cid(cell)
-
-                if cid == 0:
-                    await self._handle_tor_cmd_cell(cell)
-                else:
-                    # Replace the real circuit id by a dummy one.
-                    cell_mut = lnn.cell.header_view.write(cell, circuit_id=fake_circuit_id)
-                    #cell_mut = lnn.utils.cell_set_cid(cell, fake_circuit_id)
-
-                    self.cell_recv += 1
-                    logging.info('cell {} recv by relay: {}'.format(self.cell_recv, cell[:20].hex()))
-                    await self.channel_manager.schedule_to_send(cell_mut, cid)
-
-                (cell, data) = lnn.utils.cell_slice_old(data)
-
-            data_cells = data_cells + data
-
-            if data_cells != data_initial:
-                logging.warning('CELL SLICING MANGLE THE DATA')
-                logging.warning('INITIAL:\n{}'.format(data_initial.hex()))
-                logging.warning('CELLS:\n{}'.format(data_cells.hex()))
-                logging.warning('PREV BUFFER:\n{}'.format(self.buffer.hex()))
-                logging.warning('NEXT BUFFER:\n{}'.format(data.hex()))
-
-            # At the end we keep the beginning of the next cell.
-            self.buffer = data
-
-
-    async def _send(self, writer):
-        """
-        Send cell to the tor relay.
-        :param reader: asyncio StreamWriter
-        """
-        while not writer.is_closing():
-            cell = await self.to_send.get()
-
-            writer.write(cell)
-            await writer.drain()
-            logging.debug('Link: Sent cell: {}... {} bytes.'.format(cell[:20], len(cell)))
-
-            self.cell_sent += 1
-            logging.info('cell {} sent to relay: {}'.format(self.cell_sent, cell[:20].hex()))
-            #await asyncio.sleep(0.01)
-
-
-    async def _negociate_version(self, reader, writer, versions):
-        """
-        Negociate tor version with the relay.
-        :param reader: asyncio.StreamReader
-        :param writer: asyncio.StreamWriter
-        :param versions: collection of versions supported by the proxy
-        :return: negociated version
-        """
-
-        logging.debug('Link: Begin version negociation.')
-
-        # Ask the relay which versions it supports.
-        #payload = lnn.cell.versions.pack(versions)
-        payload = lnn.utils.cell_version_build(versions)
-
-        #await lnn.cell.versions.send_async(writer, payload)
-        # Code from this method inlined here:
-
-        #payload = payload.raw
-
-        #vercell = lnn.cell.versions.cell(payload)
-        #if not vercell.valid:
-        #    raise InvalidVersionCellException()
-
-        writer.write(payload)
-        await writer.drain()
-
-        logging.debug('Link: Sent version cell: {}.'.format(payload))
-
-        #vercell = await lnn.cell.versions.recv_async(reader)
-        # Code from this method inlined here:
-        
-        #answer = await reader.read(lnn.cell.header_legacy_view.width())
-        answer = await reader.read(5)
-
-        #header = lnn.cell.header_legacy(answer)
-        #if not header.valid:
-        #    raise InvalidCellHeaderException()
-        #if not header.cmd == lnn.cell.cmd.VERSIONS:
-        #    raise InvalidVersionCellException()
-
-        #length = header.length
-        length = lnn.utils.cell_version_get_len(answer)
-
-        if length > lnn.constants.max_payload_len:
-            raise InvalidVersionCellException()
-
-        answer += await reader.read(length)
-        #if not lnn.cell.versions.view.valid(answer):
-        #    raise InvalidVersionCellException()
-
-        logging.debug('Link: Received version cell: {}'.format(answer))
-
-        #vercell = lnn.cell.versions.cell(answer)
-        versions_res = lnn.utils.cell_version_get_versions(answer)
-
-        #common_versions = set(vercell.versions).intersection(versions)
-        common_versions = set(versions_res).intersection(versions)
-
-        if len(common_versions) < 1:
-            raise NoSupportedVersionException()
-
-        max_version = max(common_versions)
-        if max_version < 4:
-            raise NoSupportedVersionException()
-
-        # The latest version is selected.
-
-        logging.debug('Link: Select version {}.'.format(max_version))
-        logging.debug('Link: End version negociation.')
-
-        return max_version
-
-
-    async def _handler(self, host, port, ctxt, versions):
-        """
-        Establish a connection with the tor relay and handle all
-        communications between it and the proxy.
-        :param host: ip address of the relay
-        :param port: port of the relay
-        :param ctxt: TLS context
-        :param versions: collection of versions supported by the proxy
-        """
-
-        # The expected transcript is:
-        #
-        #    Onion Proxy (client)                Onion Router (server)
-        #
-        #            /   [1] :-------- VERSIONS ---------> [2]
-        #            |   [4] <-------- VERSIONS ---------: [3]
-        #            |
-        #            |           Negotiated Version
-        #            |
-        #            |   [4] <--------- CERTS -----------: [3]
-        #            |       <----- AUTH_CHALLENGE ------:
-        #            |       <-------- NETINFO ----------:
-        #            |
-        #            |             OP don't need to
-        #     Link   |               authenticate
-        #   Protocol |
-        #     >= 3   |   [5] :-------- NETINFO ----------> [6]
-        #            |
-        #            | Alternative:
-        #            | (          We (OR) authenticate         )
-        #            | (                                       )
-        #            | ( [5] :--------- CERTS -----------> [6] )
-        #            | (     :------ AUTHENTICATE ------->     )
-        #            | (             ^                         )
-        #            | (            (answers AUTH_CHALLENGE)   )
-        #            | (                                       )
-        #            \
-
-        logging.debug('Link: Opening connection.')
-
-        reader, writer = await asyncio.open_connection(host, port, ssl=ctxt)
-
-        logging.debug('Link: Connection open.')
-
-        self.version = await self._negociate_version(reader, writer, versions)
-
-        # Tor handshake
-        cells_raw = await reader.read(65536)
-
-        #logging.debug('Link: Received handshake cells:\n{}'.format(cells_raw))
-
-        (cell, cells_raw) = lnn.utils.cell_slice(cells_raw)
-        #certs_cell = lnn.cell.certs.cell(cell)
-        certs_cell = cell
-        #logging.debug('Link: Certs cell:\n{}... {} bytes.'.format(certs_cell.raw[:20], len(certs_cell.raw)))
-        logging.debug('Link: Certs cell: {}... {} bytes.'.format(certs_cell[:20], len(certs_cell)))
-
-        (cell, cells_raw) = lnn.utils.cell_slice(cells_raw)
-        #auth_cell = lnn.cell.challenge.cell(cell)
-        auth_cell = cell
-        #logging.debug('Link: Auth cell:\n{}... {} bytes.'.format(auth_cell.raw[:20], len(auth_cell.raw)))
-        logging.debug('Link: Auth cell: {}... {} bytes.'.format(auth_cell[:20], len(auth_cell)))
-
-        (cell, cells_raw) = lnn.utils.cell_slice(cells_raw)
-        #netinfo_cell = lnn.cell.netinfo.cell(cell)
-        netinfo_cell = cell
-        #logging.debug('Link: Netinfo cell:\n{}... {} bytes.'.format(netinfo_cell.raw[:20], len(netinfo_cell.raw)))
-        logging.debug('Link: Netinfo cell: {}... {} bytes.'.format(netinfo_cell[:20], len(netinfo_cell)))
-
-        # Validation of handshake cells given by the relay.
-        #if not certs_cell.valid:
-        #    raise InvalidCertsCellException()
-
-        #if not auth_cell.valid:
-        #    raise InvalidAuthCellException()
-
-        #if not netinfo_cell.valid:
-        #    raise InvalidNetInfoCellException()
-
-        # Send the NETINFO without doing any further authentication.
-        #netinfo_cell = lnn.cell.pad(lnn.cell.netinfo.pack(host))
-        netinfo_cell = lnn.utils.cell_netinfo_build(host)
-
-        writer.write(netinfo_cell)
-        await writer.drain()
-
-        logging.debug('Link: Sent netinfo cell: {}'.format(netinfo_cell))
-
-        # Handle all communication from now on.
-        tasks = [
-            asyncio.create_task(self._recv(reader)),
-            asyncio.create_task(self._send(writer))
-        ]
-
-        done, pending = await asyncio.wait(tasks, return_when=asyncio.FIRST_COMPLETED)
-
-        #while not reader.at_eof():
-        #    done, pending = await asyncio.wait(tasks, return_when=asyncio.FIRST_COMPLETED)
-        #    if tasks[0].done() or tasks[0].cancelled():
-        #        tasks[0] = asyncio.create_task(self._recv(reader))
-        #        logging.debug('Link: New recv task created.')
-        #    if tasks[1].done() or tasks[1].cancelled():
-        #        tasks[1] = asyncio.create_task(self._send(writer))
-        #        logging.debug('Link: New send task created.')
-
-        # Proper termination of communications.
-        for task in tasks:
-            if not task.cancelled():
-                task.cancel()
-
-        writer.close()
-        await writer.wait_closed()
-
-        logging.debug('Link: Connection closed.')
-
+            cell = lnn.cell.destroy.cell(payload)
+            if not cell.valid:
+                raise RuntimeError('Got invalid DESTROY cell: {}'.format(
+                    cell.truncated))
+
+            self.put(circuit, payload)
+            circuit.destroyed = True
+            circuit.reason = cell.reason
+            self.unregister(circuit)
+            logging.debug('Circuit {} got destroyed, reason: {}'.format(
+                circuit.id, circuit.reason))
+            return False
+
+        self.put(circuit, payload)
+        return True
+
+    def put(self, circuit, payload):
+        circuits_size = sum([c.queue.qsize() for _, c in self.circuits.items()])
+        if circuits_size > self.max_queue:
+            raise RuntimeError(
+                'Link circuit queues are full: {}'.format(circuits_size))
+
+        if circuit.id not in self.circuits:
+            raise RuntimeError('Got circuit_id {} outside {}, cell: {}'.format(
+                circuit.id, list(self.circuits), payload))
+
+        try:
+            payload = payload.raw
+        except AttributeError:
+            pass
+
+        self.circuits[circuit.id].put(payload)
+
+    def get(self, circuit, block=True):
+        while block and not self.io.dead:
+            try:
+                return self.circuits[circuit.id].get(block=False)
+            except queue.Empty:
+                pass
+
+            self.pull()
+            while self.pull(block=False):
+                pass
+        else:
+            if not self.io.dead:
+                while self.pull(block=False):
+                    pass
+                return self.circuits[circuit.id].get(block=False)
+            raise RuntimeError('Seems that link.io is dead!')
+
+    def register(self, circuit):
+        if circuit.id in self.circuits:
+            raise RuntimeError('Circuit {} already registered.'.format(
+                circuit.id))
+
+        circuit.queue = queue.Queue(maxsize=self.max_queue)
+        self.circuits[circuit.id] = circuit
+
+    def unregister(self, circuit):
+        del self.circuits[circuit.id]
+
+    def recv(self, block=True):
+        return self.io.recv(block=block)
+
+    def send(self, cell, block=True):
+        self.io.send(cell, block=block)
+
+    def close(self):
+        self.io.close()
+
+def negotiate_version(peer, versions, *, as_initiator):
+    """Performs a VERSIONS negotiation
+
+    :param peer: ssl.socket instance.
+    :param list versions: target link versions.
+    :param bool as_initiator: send VERSIONS cell first.
+    """
+    if as_initiator:
+        lnn.cell.versions.send(peer, lnn.cell.versions.pack(versions))
+    vercell = lnn.cell.versions.recv(peer)
+
+    common_versions = list(set(vercell.versions).intersection(versions))
+    if len(common_versions) < 1:
+        raise RuntimeError('No common supported versions: {} and {}'.format(
+            list(vercell.versions), versions))
+
+    version = max(common_versions)
+    if version < 4:
+        raise RuntimeError('No support for version 3 or lower, got {}').format(
+            version)
+
+    if not as_initiator:
+        lnn.cell.versions.send(peer, lnn.cell.versions.pack(versions))
+    return version
+
+def initiate(address='127.0.0.1', port=9050, versions=[4, 5]):
+    """Establish a link with the "in-protocol" (v3) handshake as initiator
+
+    The expected transcript is:
+
+           Onion Proxy (client)              Onion Router (server)
+
+               /   [1] :-------- VERSIONS ---------> [2]
+               |   [4] <-------- VERSIONS ---------: [3]
+               |
+               |           Negotiated Version
+               |
+               |   [4] <--------- CERTS -----------: [3]
+               |       <----- AUTH_CHALLENGE ------:
+               |       <-------- NETINFO ----------:
+               |
+               |             OP don't need to
+        Link   |               authenticate
+      Protocol |
+        >= 3   |   [5] :-------- NETINFO ----------> [6]
+               |
+               | Alternative:
+               | (          We (OR) authenticate         )
+               | (                                       )
+               | ( [5] :--------- CERTS -----------> [6] )
+               | (     :------ AUTHENTICATE ------->     )
+               | (             ^                         )
+               | (            (answers AUTH_CHALLENGE)   )
+               | (                                       )
+               \
+
+    :param str address: remote relay address (default: 127.0.0.1).
+    :param int port: remote relay ORPort (default: 9050).
+    :param list versions: target link versions (default: [4, 5]).
+
+    :returns: a link.link object
+
+    """
+
+    # Setup context
+    peer = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    ctxt = ssl.SSLContext(ssl.PROTOCOL_TLS)
+
+    # https://trac.torproject.org/projects/tor/ticket/28616
+    ctxt.options |= ssl.OP_NO_TLSv1_3
+
+    # Establish connection
+    peer = ctxt.wrap_socket(peer)
+    peer.connect((address, port))
+
+    # VERSIONS handshake
+    version = negotiate_version(peer, versions, as_initiator=True)
+
+    # Wraps with socket.io
+    peer = lnn.socket.io(peer)
+
+    # Get CERTS, AUTH_CHALLENGE and NETINFO cells afterwards
+    certs_cell = lnn.cell.certs.cell(peer.recv())
+    auth_cell = lnn.cell.challenge.cell(peer.recv())
+    netinfo_cell = lnn.cell.netinfo.cell(peer.recv())
+
+    # Sanity checks
+    if not certs_cell.valid:
+        raise RuntimeError('Invalid CERTS cell: {}'.format(certs_cell.raw))
+    if not auth_cell.valid:
+        raise RuntimeError('Invalid AUTH_CHALLENGE cell:{}'.format(
+            auth_cell.raw))
+    if not netinfo_cell.valid:
+        raise RuntimeError('Invalid NETINFO cell: {}'.format(netinfo_cell.raw))
+
+    # Send our NETINFO to say "we don't want to authenticate"
+    peer.send(lnn.cell.netinfo.pack(address))
+    return link(peer, version)
