+++ lightnion/proxy/__init__.py
@@ -3,4 +3,4 @@ fake_circuit_id = 0x80000000
 api_version = 0.1
 base_url = '/lightnion/api/v{}'.format(api_version)
 
-from . import path, parts, auth, jobs
+from . import parts, auth, jobs, link
